Sube logica de manejo de excepciones y logica de recupero del ultimo porcentaje obtenido.

+++ challenge/src/main/java/com/tenpo/challenge/service/CalculatedRecordServiceImpl.java
@@ -1,11 +1,15 @@
 package com.tenpo.challenge.service;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.logging.Logger;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 
 import com.tenpo.challenge.model.CalculatedRecord;
@@ -35,21 +39,40 @@ public class CalculatedRecordServiceImpl {
 	}
 	
 	/**
+	 * Metodo que recupera el ultimo porcentaje que vino en el servicio externo.
+	 * @return ultimo porcentaje utilizado
+	 * @throws Exception en caso que no haya registros en la BD.
+	 */
+	private int getLastPercentage() throws Exception{
+		
+		List<CalculatedRecord> registros = calculatedRecordRepository.findAll(
+				Sort.by(Direction.DESC, "creationDttm"));
+		
+		if(registros.isEmpty()) {
+			throw new Exception("No hay ningún valor configurado para el porcentaje en la BD.");
+		}
+		
+		return registros.get(0).getValue_external_service();
+	}
+	
+	/**
 	 * Metodo que realiza el calculo y la persistencia en la BD.
 	 * @param record
 	 * @param thirdPartyPercentage 
 	 * @return
+	 * @throws Exception 
 	 */
-	public CalculatedRecord createOrUpdateRecord(CalculatedRecord record, ThirdPartyPercentage thirdPartyPercentage) {
+	public CalculatedRecord createOrUpdateRecord(CalculatedRecord record, ThirdPartyPercentage thirdPartyPercentage) throws Exception {
 		
-		//invocar al servicio externo para recuperar el %
-		int porcentaje = thirdPartyPercentage!=null?thirdPartyPercentage.getPorcentage():0;
+		//invocar al servicio externo para recuperar el %,
+		//en caso que no lo encuentre, recupera el ultimo
+		int porcentaje = thirdPartyPercentage!=null?thirdPartyPercentage.getPorcentage():getLastPercentage();
 		
 		record.setValue_external_service(porcentaje);
 		record.setResult(record.getValue_1() + record.getValue_2() +
 				calcularPorcentaje(record.getValue_1() + record.getValue_2(), porcentaje));
 		
-		//enviaa el mensaje a rabbit
+		//envia el mensaje a rabbit
 		sentToRabbit(record);
 		
 		return record;
@@ -73,13 +96,13 @@ public class CalculatedRecordServiceImpl {
 	 * @return
 	 */
 	public boolean deleteById(Long id) {
-		try {
-			calculatedRecordRepository.deleteById(id);
-			return true;
-		}catch (Exception e) {
-			e.printStackTrace();
-			return false;
+
+		if(!findById(id).isPresent()) {
+			throw new EmptyResultDataAccessException(1);
 		}
+		
+		calculatedRecordRepository.deleteById(id);
+		return true;
 	}
 	
 	/**
@@ -88,9 +111,20 @@ public class CalculatedRecordServiceImpl {
 	 * @return
 	 */
 	public Optional<CalculatedRecord> findById(Long id){
-		return calculatedRecordRepository.findById(id);
+		
+		Optional<CalculatedRecord> record = calculatedRecordRepository.findById(id);
+		
+		if(!record.isPresent()) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
+		return record;
 	}
 	
+	/**
+	 * Envia el mensaje a rabbit.
+	 * @param message
+	 */
 	private void sentToRabbit(CalculatedRecord message) {
 		
 		LOGGER.info("Mensaje { " + message.toString() + " } sera enviado...");
